app/requirements.txt
Flask
uwsgi
Flask-SQLAlchemy
Flask-Login
Flask-Bcrypt
CommonMark
SQLAlchemy-migrate
psycopg2
